check for updates via the slippi logo
