+++ src/renderer/components/Footer.tsx
@@ -3,15 +3,10 @@ import { slippiDonateUrl, socials } from "@common/constants";
 import { css } from "@emotion/react";
 import styled from "@emotion/styled";
 import TwitterIcon from "@mui/icons-material/Twitter";
-import { Button } from "@mui/material";
 import Tooltip from "@mui/material/Tooltip";
 import React from "react";
 
 import { ExternalLink as A } from "@/components/ExternalLink";
-import { useDolphinActions } from "@/lib/dolphin/useDolphinActions";
-import { useAppUpdate } from "@/lib/hooks/useAppUpdate";
-import { useToasts } from "@/lib/hooks/useToasts";
-import { useServices } from "@/services";
 import { ReactComponent as DiscordIcon } from "@/styles/images/discord.svg";
 
 export const BasicFooter = styled.div`
@@ -31,40 +26,8 @@ export const BasicFooter = styled.div`
 `;
 
 export const Footer: React.FC = () => {
-  const { showInfo, showError } = useToasts();
-  const { dolphinService } = useServices();
-  const { checkForAppUpdates } = useAppUpdate();
-  const { updateDolphin } = useDolphinActions(dolphinService);
-  const [checkingForUpdates, setCheckingForUpdates] = React.useState(false);
-  const checkForUpdatesHandler = React.useCallback(async () => {
-    setCheckingForUpdates(true);
-    try {
-      showInfo("Checking for updates...");
-      await checkForAppUpdates();
-      await updateDolphin();
-    } catch (err) {
-      window.electron.log.error(err);
-      showError("Failed to get updates");
-    }
-    setTimeout(() => setCheckingForUpdates(false), 10000);
-  }, [checkForAppUpdates, updateDolphin, showInfo, showError]);
-
   return (
     <Outer>
-      <Button
-        css={css`
-          text-transform: uppercase;
-          font-weight: bold;
-          font-size: 12px;
-          margin-right: auto;
-          justify-content: flex-start;
-          color: ${colors.purpleLight};
-        `}
-        onClick={checkForUpdatesHandler}
-        disabled={checkingForUpdates}
-      >
-        Check for updates
-      </Button>
       <Social title="Follow Project Slippi on Twitter" url={`https://twitter.com/${socials.twitterId}`}>
         <TwitterIcon />
       </Social>
